+++ fabric/src/main/java/io/github/mortuusars/envelope/fabric/EnvelopeFabricClient.java
@@ -7,12 +7,10 @@ import io.github.mortuusars.envelope.client.gui.screen.PackingScreen;
 import io.github.mortuusars.envelope.client.gui.screen.PaybackPackingScreen;
 import io.github.mortuusars.envelope.client.gui.screen.PaybackTagScreen;
 import io.github.mortuusars.envelope.client.gui.screen.MailboxScreen;
-import io.github.mortuusars.envelope.client.model.PigeonBackpackModel;
-import io.github.mortuusars.envelope.client.model.PigeonFancyHatModel;
-import io.github.mortuusars.envelope.client.model.PigeonLegBandModel;
 import io.github.mortuusars.envelope.client.model.PigeonModel;
-import io.github.mortuusars.envelope.client.model.geom.EnvelopeModelLayers;
 import io.github.mortuusars.envelope.client.renderer.entity.PigeonRenderer;
+import io.github.mortuusars.envelope.client.renderer.entity.layer.PigeonBackpackLayer;
+import io.github.mortuusars.envelope.client.renderer.entity.layer.PigeonHatLayer;
 import io.github.mortuusars.envelope.network.fabric.FabricS2CPacketHandler;
 import io.github.mortuusars.envelope.world.item.Sealable;
 import net.fabricmc.api.ClientModInitializer;
@@ -30,10 +28,9 @@ public class EnvelopeFabricClient implements ClientModInitializer {
         ConfigScreenFactoryRegistry.INSTANCE.register(Envelope.ID, ConfigurationScreen::new);
 
         EntityRendererRegistry.register(Envelope.EntityTypes.PIGEON.get(), PigeonRenderer::new);
-        EntityModelLayerRegistry.registerModelLayer(EnvelopeModelLayers.PIGEON, PigeonModel::createLayerDefinition);
-        EntityModelLayerRegistry.registerModelLayer(EnvelopeModelLayers.PIGEON_LEG_BAND, PigeonLegBandModel::createLayerDefinition);
-        EntityModelLayerRegistry.registerModelLayer(EnvelopeModelLayers.PIGEON_BACKPACK, PigeonBackpackModel::createLayerDefinition);
-        EntityModelLayerRegistry.registerModelLayer(EnvelopeModelLayers.PIGEON_FANCY_HAT, PigeonFancyHatModel::createLayerDefinition);
+        EntityModelLayerRegistry.registerModelLayer(PigeonRenderer.PIGEON_LAYER, PigeonModel::createLayerDefinition);
+        EntityModelLayerRegistry.registerModelLayer(PigeonBackpackLayer.PIGEON_BACKPACK, PigeonModel::createLayerDefinition);
+        EntityModelLayerRegistry.registerModelLayer(PigeonHatLayer.PIGEON_HAT, PigeonModel::createLayerDefinition);
 
         ColorProviderRegistry.ITEM.register(Sealable::getSealOverlayColor, Envelope.Items.SEALED_LETTER.get());
         ColorProviderRegistry.ITEM.register(Sealable::getSealOverlayColor, Envelope.Items.SEALED_PACKAGE.get());